+++ src/main/java/com/searchitemsapp/entities/TbSiaSelectoresCss.java
@@ -26,7 +26,7 @@ public class TbSiaSelectoresCss implements Serializable {
 
 	@Id
 	@Column(name = "did")
-	private Integer did;
+	private Long did;
 
 	@Column(name="bol_activo")
 	private Boolean bolActivo;
@@ -68,11 +68,11 @@ public class TbSiaSelectoresCss implements Serializable {
 	@JoinColumn(name="did_url", referencedColumnName="did", nullable = false)
 	private TbSiaUrl tbSiaUrl;
 
-	public Integer getDid() {
+	public Long getDid() {
 		return this.did;
 	}
 
-	public void setDid(Integer did) {
+	public void setDid(Long did) {
 		this.did = did;
 	}
