+++ src/main/java/com/searchitemsapp/entities/TbSiaUrl.java
@@ -29,7 +29,7 @@ public class TbSiaUrl implements Serializable {
 
 	@Id
 	@Column(name = "did")
-	private Integer did;
+	private Long did;
 
 	@Column(name="bol_activo")
 	private Boolean bolActivo;
@@ -56,11 +56,11 @@ public class TbSiaUrl implements Serializable {
 	@OneToMany(mappedBy="tbSiaUrl")
 	private List<TbSiaSelectoresCss> tbSiaSelectoresCsses;	
 	
-	public Integer getDid() {
+	public Long getDid() {
 		return this.did;
 	}
 
-	public void setDid(Integer did) {
+	public void setDid(Long did) {
 		this.did = did;
 	}
